🐛 fix int-map key types
We currently don't have deep support for int-maps, since we don't have
any use-cases for them yet. That said, they have been supported in the
underlying stack for when we need them. Unfortunately a few of the
supporting elements have been out of sync.

This change aligns data conversions in llx to use int64 maps. Ditto for
`isTruthy` in rawdata. There are a few more to tackle...

+++ llx/rawdata.go
@@ -322,7 +322,7 @@ func isTruthy(data interface{}, typ types.Type) (bool, bool) {
 			}
 
 		case types.Int:
-			m := data.(map[int]interface{})
+			m := data.(map[int64]interface{})
 			for _, v := range m {
 				t1, f1 := isTruthy(v, typ.Child())
 				if f1 {
